@@ -2,46 +2,52 @@
     "name": "totally",
     "version": "2.0.0",
     "private": true,
+    "packageManager": "pnpm@9.14.4",
+    "engines": {
+        "npm": "please-use-pnpm",
+        "yarn": "please-use-pnpm"
+    },
     "scripts": {
-        "dev": "next dev --turbo",
+        "preinstall": "npx only-allow pnpm",
+        "dev": "next dev",
         "build": "next build",
         "start": "next start",
         "lint": "next lint"
     },
     "dependencies": {
-        "@radix-ui/react-alert-dialog": "^1.1.1",
-        "@radix-ui/react-avatar": "^1.1.0",
-        "@radix-ui/react-dialog": "^1.1.1",
-        "@radix-ui/react-dropdown-menu": "^2.1.1",
-        "@radix-ui/react-icons": "^1.3.0",
+        "@radix-ui/react-alert-dialog": "^1.1.2",
+        "@radix-ui/react-avatar": "^1.1.1",
+        "@radix-ui/react-dialog": "^1.1.2",
+        "@radix-ui/react-dropdown-menu": "^2.1.2",
+        "@radix-ui/react-icons": "^1.3.2",
         "@radix-ui/react-label": "^2.1.0",
-        "@radix-ui/react-menubar": "^1.1.1",
+        "@radix-ui/react-menubar": "^1.1.2",
         "@radix-ui/react-slot": "^1.1.0",
-        "@tanstack/react-table": "^8.20.1",
-        "babel-plugin-react-compiler": "^0.0.0-experimental-334f00b-20240725",
-        "class-variance-authority": "^0.7.0",
+        "@tanstack/react-table": "^8.20.5",
+        "babel-plugin-react-compiler": "^19.0.0-beta-df7b47d-20241124",
+        "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
-        "lucide-react": "^0.424.0",
-        "next": "^15.0.0-canary.102",
-        "next-themes": "^0.3.0",
+        "jotai": "^2.10.3",
+        "lucide-react": "^0.462.0",
+        "next": "^15.0.3",
+        "next-themes": "^0.4.3",
         "random-words": "^2.0.1",
-        "react": "^19.0.0-rc-8269d55d-20240802",
-        "react-dom": "^19.0.0-rc-8269d55d-20240802",
-        "tailwind-merge": "^2.4.0",
-        "tailwindcss-animate": "^1.0.7",
-        "zustand": "^4.5.4"
+        "react": "^19.0.0-rc-70fb1363-20241010",
+        "react-dom": "^19.0.0-rc-70fb1363-20241010",
+        "tailwind-merge": "^2.5.5",
+        "tailwindcss-animate": "^1.0.7"
     },
     "devDependencies": {
         "@types/node": "^22",
         "@types/react": "^18",
         "@types/react-dom": "^18",
         "autoprefixer": "^10.4.20",
-        "eslint": "^8.57.0",
-        "eslint-config-next": "14.2.5",
+        "eslint": "^9.16.0",
+        "eslint-config-next": "15.0.3",
         "eslint-config-prettier": "^9.1.0",
         "postcss": "^8",
-        "prettier": "3.3.3",
-        "tailwindcss": "^3.4.7",
-        "typescript": "^5.5.4"
+        "prettier": "3.4.1",
+        "tailwindcss": "^3.4.15",
+        "typescript": "^5.7.2"
     }
 }
