@@ -1,5 +1,7 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
+import type { NextConfig } from "next"
+
+const nextConfig: NextConfig = {
+    reactStrictMode: true,
     images: {
         dangerouslyAllowSVG: true,
         domains: ["api.dicebear.com"],
@@ -13,8 +15,8 @@ const nextConfig = {
         ]
     },
     experimental: {
-        reactCompiler: true
+        reactCompiler: false
     }
 }
 
-module.exports = nextConfig
+export default nextConfig
