@@ -1,12 +1,7 @@
 "use client"
 
+import Scoreboard from "@/app/(scoreboard)/scoreboard"
 import Title from "@/components/title"
-import { usePlayerStore } from "@/lib/stores/players"
-import dynamic from "next/dynamic"
-
-const Scoreboard = dynamic(() => import("@/components/scoreboard"), {
-    ssr: false
-})
 
 export default function ScoreboardPage() {
     return (
