@@ -1,15 +1,21 @@
 import Title from "@/components/title"
 import dynamic from "next/dynamic"
 
-const Round = dynamic(() => import("@/components/round"), {
-    ssr: false
-})
+const Round = dynamic(() => import("@/components/round"))
 
-export default function RoundPage({ params }: { params: { round: string } }) {
+type Params = Promise<{ round: string }>
+
+export default async function RoundPage({ params }: { params: Params }) {
+    const { round } = await params
+
+    if (isNaN(Number(round))) {
+        return <div>Invalid round number</div>
+    }
+
     return (
         <>
-            <Title>Round {params.round}</Title>
-            <Round round={Number(params.round)} />
+            <Title>Round {round}</Title>
+            <Round round={Number(round)} />
         </>
     )
 }
