@@ -1,13 +1,15 @@
 "use client"
 
-import { useSharingStore } from "@/lib/stores/sharing"
+import { useAtom, useAtomValue } from "jotai"
+import { uniqueAppCodeAtom } from "@/lib/atoms/players"
+import { generate } from "random-words"
 
 const AppCode = () => {
-    const appCode = useSharingStore((state) => state.uniqueAppCode)
-    const reset = useSharingStore((state) => state.resetUniqueAppCode)
+    const appCode = useAtomValue(uniqueAppCodeAtom)
+    const [, setAppCode] = useAtom(uniqueAppCodeAtom)
 
     if (!appCode || appCode === "") {
-        reset()
+        setAppCode(generate({ minLength: 4 }) as string)
     }
 
     return <div className="text-2xl font-semibold p-3 font-mono text-center">{appCode}</div>
